+++ src/main/kotlin/backend/itracker/schedule/service/CrawlSchedulerService.kt
@@ -1,17 +1,18 @@
 package backend.itracker.schedule.service
 
 import backend.itracker.crawl.service.CrawlTargetCategory
-import backend.itracker.schedule.service.crawlers.CrawlComposite
+import backend.itracker.schedule.service.schedulers.SchedulerComposite
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Component
 
 private const val CRAWLING_TIME = "30 31 17 * * *"
 private const val TIME_ZONE = "Asia/Seoul"
+
 private val CRAWL_TARGETS = CrawlTargetCategory.entries
 
 @Component
 class CrawlSchedulerService(
-    private val crawlers: CrawlComposite,
+    private val crawlers: SchedulerComposite,
 ) {
 
     fun crawlManual(category: CrawlTargetCategory) {
