+++ src/main/kotlin/backend/itracker/schedule/service/schedulers/airpods/AirpodsScheduler.kt
@@ -1,10 +1,10 @@
-package backend.itracker.schedule.service.crawlers.macbook
+package backend.itracker.schedule.service.schedulers.airpods
 
-import backend.itracker.crawl.macbook.service.MacbookService
+import backend.itracker.crawl.airpods.service.AirPodsService
 import backend.itracker.crawl.service.CrawlResultService
 import backend.itracker.crawl.service.CrawlService
 import backend.itracker.crawl.service.CrawlTargetCategory
-import backend.itracker.schedule.service.crawlers.Schedulable
+import backend.itracker.schedule.service.schedulers.Schedulable
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.springframework.stereotype.Component
 import kotlin.time.measureTime
@@ -12,23 +12,23 @@ import kotlin.time.measureTime
 private val logger = KotlinLogging.logger {}
 
 @Component
-class MacbookScheduler(
-    private val macbookService: MacbookService,
+class AirpodsScheduler(
+    private val airPodsService: AirPodsService,
     private val crawlResultService: CrawlResultService,
     private val crawlService: CrawlService,
 ) : Schedulable(crawlResultService, crawlService) {
 
     override fun supports(): CrawlTargetCategory {
-        return CrawlTargetCategory.MACBOOK
+        return CrawlTargetCategory.AIRPODS
     }
 
     override fun doSchedule() {
-        logger.info { "맥북 크롤링 시작. " }
+        logger.info { "에어팟 크롤링 시작. " }
         val times = measureTime {
-            val macbooks = crawlService.crawlMacbook()
-            macbookService.saveAll(macbooks)
+            val airPods = crawlService.crawlAirPods()
+            airPodsService.saveAll(airPods)
         }
-        crawlResultService.updateCrawlResult(CrawlTargetCategory.MACBOOK)
-        logger.info { "맥북 크롤링 끝. 시간: $times" }
+        crawlResultService.updateCrawlResult(CrawlTargetCategory.AIRPODS)
+        logger.info { "에어팟 크롤링 끝. 시간: $times" }
     }
 }
