+++ src/main/kotlin/backend/itracker/schedule/service/schedulers/ipad/IpadScheduler.kt
@@ -1,10 +1,10 @@
-package backend.itracker.schedule.service.crawlers.mac
+package backend.itracker.schedule.service.schedulers.ipad
 
-import backend.itracker.crawl.mac.service.MacService
+import backend.itracker.crawl.ipad.service.IpadService
 import backend.itracker.crawl.service.CrawlResultService
 import backend.itracker.crawl.service.CrawlService
 import backend.itracker.crawl.service.CrawlTargetCategory
-import backend.itracker.schedule.service.crawlers.Schedulable
+import backend.itracker.schedule.service.schedulers.Schedulable
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.springframework.stereotype.Component
 import kotlin.time.measureTime
@@ -12,23 +12,23 @@ import kotlin.time.measureTime
 private val logger = KotlinLogging.logger {}
 
 @Component
-class MacScheduler(
-    private val macService: MacService,
+class IpadScheduler(
+    private val ipadService: IpadService,
     private val crawlResultService: CrawlResultService,
     private val crawlService: CrawlService,
 ) : Schedulable(crawlResultService, crawlService) {
 
     override fun supports(): CrawlTargetCategory {
-        return CrawlTargetCategory.MAC
+        return CrawlTargetCategory.IPAD
     }
 
     override fun doSchedule() {
-        logger.info { "맥 크롤링 시작. " }
+        logger.info { "아이패드 크롤링 시작. " }
         val times = measureTime {
-            val macs = crawlService.crawlMac()
-            macService.saveAll(macs)
+            val ipads = crawlService.crawlIpad()
+            ipadService.saveAll(ipads)
         }
-        crawlResultService.updateCrawlResult(CrawlTargetCategory.MAC)
-        logger.info { "맥 크롤링 끝. 시간: $times" }
+        crawlResultService.updateCrawlResult(CrawlTargetCategory.IPAD)
+        logger.info { "아이패드 크롤링 끝. 시간: $times" }
     }
 }
