+++ src/main/kotlin/backend/itracker/schedule/service/schedulers/mac/MacScheduler.kt
@@ -1,10 +1,10 @@
-package backend.itracker.schedule.service.crawlers.watch
+package backend.itracker.schedule.service.schedulers.mac
 
+import backend.itracker.crawl.mac.service.MacService
 import backend.itracker.crawl.service.CrawlResultService
 import backend.itracker.crawl.service.CrawlService
 import backend.itracker.crawl.service.CrawlTargetCategory
-import backend.itracker.crawl.watch.service.AppleWatchService
-import backend.itracker.schedule.service.crawlers.Schedulable
+import backend.itracker.schedule.service.schedulers.Schedulable
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.springframework.stereotype.Component
 import kotlin.time.measureTime
@@ -12,23 +12,23 @@ import kotlin.time.measureTime
 private val logger = KotlinLogging.logger {}
 
 @Component
-class AppleWatchScheduler(
-    private val appleWatchService: AppleWatchService,
+class MacScheduler(
+    private val macService: MacService,
     private val crawlResultService: CrawlResultService,
     private val crawlService: CrawlService,
 ) : Schedulable(crawlResultService, crawlService) {
 
     override fun supports(): CrawlTargetCategory {
-        return CrawlTargetCategory.APPLE_WATCH
+        return CrawlTargetCategory.MAC
     }
 
     override fun doSchedule() {
-        logger.info { "애플워치 크롤링 시작. " }
+        logger.info { "맥 크롤링 시작. " }
         val times = measureTime {
-            val appleWatches = crawlService.crawlAppleWatch()
-            appleWatchService.saveAll(appleWatches)
+            val macs = crawlService.crawlMac()
+            macService.saveAll(macs)
         }
-        crawlResultService.updateCrawlResult(CrawlTargetCategory.APPLE_WATCH)
-        logger.info { "애플워치 크롤링 끝. 시간: $times" }
+        crawlResultService.updateCrawlResult(CrawlTargetCategory.MAC)
+        logger.info { "맥 크롤링 끝. 시간: $times" }
     }
 }
