+++ src/main/kotlin/backend/itracker/schedule/service/schedulers/macbook/MacbookScheduler.kt
@@ -1,10 +1,10 @@
-package backend.itracker.schedule.service.crawlers.ipad
+package backend.itracker.schedule.service.schedulers.macbook
 
-import backend.itracker.crawl.ipad.service.IpadService
+import backend.itracker.crawl.macbook.service.MacbookService
 import backend.itracker.crawl.service.CrawlResultService
 import backend.itracker.crawl.service.CrawlService
 import backend.itracker.crawl.service.CrawlTargetCategory
-import backend.itracker.schedule.service.crawlers.Schedulable
+import backend.itracker.schedule.service.schedulers.Schedulable
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.springframework.stereotype.Component
 import kotlin.time.measureTime
@@ -12,23 +12,23 @@ import kotlin.time.measureTime
 private val logger = KotlinLogging.logger {}
 
 @Component
-class IpadScheduler(
-    private val ipadService: IpadService,
+class MacbookScheduler(
+    private val macbookService: MacbookService,
     private val crawlResultService: CrawlResultService,
     private val crawlService: CrawlService,
 ) : Schedulable(crawlResultService, crawlService) {
 
     override fun supports(): CrawlTargetCategory {
-        return CrawlTargetCategory.IPAD
+        return CrawlTargetCategory.MACBOOK
     }
 
     override fun doSchedule() {
-        logger.info { "아이패드 크롤링 시작. " }
+        logger.info { "맥북 크롤링 시작. " }
         val times = measureTime {
-            val ipads = crawlService.crawlIpad()
-            ipadService.saveAll(ipads)
+            val macbooks = crawlService.crawlMacbook()
+            macbookService.saveAll(macbooks)
         }
-        crawlResultService.updateCrawlResult(CrawlTargetCategory.IPAD)
-        logger.info { "아이패드 크롤링 끝. 시간: $times" }
+        crawlResultService.updateCrawlResult(CrawlTargetCategory.MACBOOK)
+        logger.info { "맥북 크롤링 끝. 시간: $times" }
     }
 }
